refactor(web): retire the legacy google authorization shim
#3435 parameterized sign-in start and callback by provider but left
packages/web/src/auth/google/authorization/ behind as pass-through
wrappers, so the codebase carried two parallel seams for one flow:
two test registries for the same hook (google + provider), a
google-specific availability API aliased to the neutral one, and a
callback test living in a directory whose component had moved.

Delete the shim and point the remaining callers at the provider-neutral
modules. Coverage that only existed on the shim moves onto the neutral
modules rather than being dropped: storage edge cases become
provider-authorization.storage.test.ts (previously untested), and the
callback-url, safe-return-path and auth-code-payload cases join
provider-authorization.util.test.ts.

Behavior preserving: storage keys, error copy and the google consent
retry are unchanged.
