--- packages/web/src/auth/google/authorization/google-authorization.constants.ts
@@ -1,4 +1,0 @@
-import { GOOGLE_SCOPES } from "@core/providers/google.scopes";
-
-export const GOOGLE_AUTH_SCOPES_REQUIRED = [...GOOGLE_SCOPES];
-export const GOOGLE_AUTH_SCOPES_REQUESTED = GOOGLE_AUTH_SCOPES_REQUIRED;
