+++ packages/web/src/auth/providers/authorization/provider-authorization.util.test.ts
@@ -2,8 +2,76 @@ import { MICROSOFT_SCOPES } from "@core/providers/microsoft.scopes";
 import {
   buildMicrosoftAuthorizationUrl,
   buildProviderAuthCallbackUrl,
+  buildProviderAuthCodePayload,
+  getSafeProviderAuthReturnPath,
 } from "./provider-authorization.util";
 import { describe, expect, it } from "bun:test";
+
+describe("buildProviderAuthCallbackUrl", () => {
+  it("builds the callback URL from the current origin", () => {
+    expect(
+      buildProviderAuthCallbackUrl("google", "http://localhost:9080"),
+    ).toBe("http://localhost:9080/auth/google/callback");
+  });
+});
+
+describe("getSafeProviderAuthReturnPath", () => {
+  it("keeps same-origin app paths as return paths", () => {
+    expect(
+      getSafeProviderAuthReturnPath(
+        "google",
+        "http://localhost:9080/day/2026-05-05?x=1#agenda",
+        "http://localhost:9080",
+      ),
+    ).toBe("/day/2026-05-05?x=1#agenda");
+  });
+
+  it("strips transient auth-modal params but keeps other params", () => {
+    expect(
+      getSafeProviderAuthReturnPath(
+        "google",
+        "http://localhost:9080/week/2026-07-08?auth=login&token=abc&x=1",
+        "http://localhost:9080",
+      ),
+    ).toBe("/week/2026-07-08?x=1");
+  });
+
+  it("falls back to /week for external return paths", () => {
+    expect(
+      getSafeProviderAuthReturnPath(
+        "google",
+        "https://evil.example/phish",
+        "http://localhost:9080",
+      ),
+    ).toBe("/week");
+  });
+});
+
+describe("buildProviderAuthCodePayload", () => {
+  it("builds the existing auth-code payload shape", () => {
+    expect(
+      buildProviderAuthCodePayload({
+        provider: "google",
+        code: "auth-code",
+        scope: "email profile",
+        state: "state-1",
+        redirectUri: "http://localhost:9080/auth/google/callback",
+      }),
+    ).toEqual({
+      thirdPartyId: "google",
+      clientType: "web",
+      redirectURIInfo: {
+        redirectURIOnProviderDashboard:
+          "http://localhost:9080/auth/google/callback",
+        redirectURIQueryParams: {
+          code: "auth-code",
+          scope: "email profile",
+          state: "state-1",
+        },
+      },
+    });
+  });
+});
 
 describe("buildMicrosoftAuthorizationUrl", () => {
   it("builds the common Microsoft authorize URL", () => {
